# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API GEST_RV_CHUN",
        "description": "Documentation des endpoints API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur courant"
        }
    ],
    "paths": {
...
        "/journal": {
            "get": {
                "tags": [
                    "Journal Audit"
                ],
                "summary": "Lister les journaux d'audit",
                "operationId": "1c2cec88c75a48703c28fef96cc5716c",
                "parameters": [
                    {
                        "name": "type_action",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personel_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "adresse_ip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 200
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des journaux d'audit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JournalAuditListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/journal/{id}": {
            "get": {
                "tags": [
                    "Journal Audit"
                ],
                "summary": "Afficher le détail d'un journal d'audit",
                "operationId": "67e162ee792b06ad3a5d415945471bc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du journal d'audit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JournalAuditResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Journal introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/journal/export": {
            "get": {
                "tags": [
                    "Journal Audit"
                ],
                "summary": "Exporter les journaux d'audit en CSV",
                "operationId": "1ecdbe9955f6046e60a3295d3906de3c",
                "parameters": [
                    {
                        "name": "type_action",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personel_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_debut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_fin",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "adresse_ip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier CSV généré",
                        "content": {
                            "text/csv": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 8
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "/api/v1/medecins?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "example": "/api/v1/medecins?page=8",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "/api/v1/medecins?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifie"
                    },
                    "data": {
                        "nullable": true
                    },
                    "errors": {
                        "properties": {
                            "type": {
                                "type": "string",
                                "example": "authentication"
                            },
                            "detail": {
                                "type": "string",
                                "example": "Authentification requise pour acceder a cette ressource."
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "JournalAuditAuteur": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "9f8c9c2b-2d2a-4e75-b1a6-1a2f1e2a3b4c"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Diop"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Awa"
                    },
                    "matricule": {
                        "type": "string",
                        "example": "MED-0001"
                    },
                    "role": {
                        "type": "string",
                        "example": "ADMIN"
                    }
                },
                "type": "object",
                "nullable": true
            },
            "JournalAudit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 125
                    },
                    "type_action": {
                        "type": "string",
                        "example": "CREATIONRV"
                    },
                    "details": {
                        "type": "object",
                        "example": {
                            "matricule_patient": "PAT-FANN-2026-0001",
                            "nom_patient": "Moussa Ba",
                            "service_nom": "Cardiologie",
                            "date_rendez_vous": "2026-04-10",
                            "heure_approximative": "09:00:00",
                            "medecin_nom": "Awa Diop"
                        },
                        "additionalProperties": true
                    },
                    "adresse_ip": {
                        "type": "string",
                        "example": "127.0.0.1",
                        "nullable": true
                    },
                    "user_agent": {
                        "type": "string",
                        "example": "Mozilla/5.0",
                        "nullable": true
                    },
                    "auteur": {
                        "$ref": "#/components/schemas/JournalAuditAuteur"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "25/03/2026 14:30:00"
                    }
                },
                "type": "object"
            },
            "PaginatedJournalAudits": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JournalAudit"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    }
                },
                "type": "object"
            },
            "JournalAuditResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Détail du journal d'audit"
                    },
                    "data": {
                        "$ref": "#/components/schemas/JournalAudit"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "JournalAuditListResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Liste des journaux d'audit"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedJournalAudits"
                    },
                    "errors": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}